DRY up the logout controllers and fix failing specs
