+++ app/controllers/stormpath/rails/id_site_logout/create_controller.rb
@@ -1,7 +1,7 @@
 module Stormpath
   module Rails
-    module IdSite
-      class LogoutController < BaseController
+    module IdSiteLogout
+      class CreateController < BaseController
         def call
           begin
             payload = { 'iat' => Time.now.to_i,
@@ -13,30 +13,10 @@ module Stormpath
                         'state' => '' }
             secret = ENV['STORMPATH_API_KEY_SECRET']
             jwt = JWT.encode(payload, secret, 'HS256')
-            delete_tokens
-            delete_cookies
+            TokenAndCookiesCleaner.new(cookies).remove
             redirect_to "https://api.stormpath.com/sso/logout?jwtRequest=#{jwt}"
           rescue Stormpath::Error, LoginForm::FormError => error
           end
-        end
-
-        private
-        def delete_tokens
-          DeleteAccessToken.call(cookies[access_token_cookie_name])
-          DeleteRefreshToken.call(cookies[refresh_token_cookie_name])
-        end
-
-        def delete_cookies
-          cookies.delete(access_token_cookie_name)
-          cookies.delete(refresh_token_cookie_name)
-        end
-
-        def access_token_cookie_name
-          stormpath_config.web.access_token_cookie.name
-        end
-
-        def refresh_token_cookie_name
-          stormpath_config.web.refresh_token_cookie.name
         end
       end
     end
