+++ spec/dummy/config/stormpath.yml
@@ -179,7 +179,7 @@ stormpath:
     # login, registration, and password reset.  They should also be redirected
     # through ID Site on logout.
     idSite:
-      enabled: true
+      enabled: false
       loginUri: ""
       forgotUri: "/#/forgot"
       registerUri: "/#/register"
